# src/docs/swagger.yaml
basePath: /
definitions:
...
  httputil.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
...
  model.Category:
    properties:
      category:
        type: string
      id:
        type: integer
    type: object
...
  model.CustomerAnalytic:
    properties:
      amountSpent:
        type: number
      packsPurchased:
        type: integer
      username:
        type: string
    type: object
...
  model.PackAnalyticsResp:
    properties:
      dataSet:
        items:
          $ref: '#/definitions/model.PackSalesAnalytic'
        type: array
      timeAxis:
        items:
          type: string
        type: array
    type: object
...
  model.PackQtySoldAnalytic:
    properties:
      packTitle:
        type: string
      qtySold:
        type: integer
    type: object
  model.PackSalesAnalytic:
    properties:
      granularity:
        type: string
      qtySold:
        type: integer
      totalSales:
        type: number
    type: object
...
host: apidev.xopacks.com
info:
  contact: {}
  description: The official XOPAcks API
  title: XOPacks API
  version: "1.0"
paths:
...
  /analytics/avgPacksPurchased/:vendorId:
    get:
      consumes:
      - application/json
      description: Get the average amount of packs purchased by a user at a time
      parameters:
      - description: vendor uid
        in: path
        name: vendorId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Get average pack qty purchased
      tags:
      - Analytics
  /analytics/customers/:vendorId:
    get:
      consumes:
      - application/json
      description: Get a list of the top customers for a specific vendor
      parameters:
      - description: vendor uid
        in: path
        name: vendorId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.CustomerAnalytic'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Get list of top customers
      tags:
      - Analytics
  /analytics/favoriteAmount/:vendorId:
    get:
      consumes:
      - application/json
      description: Get the current favorite amount for this vendor
      parameters:
      - description: vendor uid
        in: path
        name: vendorId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Get the favorite amount for vendor
      tags:
      - Analytics
  /analytics/maxPacksPurchased/:vendorId:
    get:
      consumes:
      - application/json
      description: Get the max amount of packs purchased by a user at a time
      parameters:
      - description: vendor uid
        in: path
        name: vendorId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Get the max pack qty purchased
      tags:
      - Analytics
  /analytics/minPacksPurchased/:vendorId:
    get:
      consumes:
      - application/json
      description: Get the min amount of packs purchased by a user at a time
      parameters:
      - description: vendor uid
        in: path
        name: vendorId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Get the min pack qty purchased
      tags:
      - Analytics
  /analytics/packQtySold/:vendorId:
    get:
      consumes:
      - application/json
      description: Get a list of packs and their qty sold
      parameters:
      - description: vendor uid
        in: path
        name: vendorId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.PackQtySoldAnalytic'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Get the pack qty sold analytics
      tags:
      - Analytics
  /analytics/packSales/:vendorId:
    get:
      consumes:
      - application/json
      description: Get pack sales analytics for a vendor
      parameters:
      - description: vendor uid
        in: path
        name: vendorId
        required: true
        type: string
      - description: time granularity
        in: query
        name: granularity
        required: true
        type: string
      - description: month filter
        in: query
        name: month
        required: true
        type: integer
      - description: year filter
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PackAnalyticsResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Get the pack sales analytics
      tags:
      - Analytics
  /analytics/totalPacksSold/:vendorId:
    get:
      consumes:
      - application/json
      description: Get a total value of all packs sold by this vendor
      parameters:
      - description: vendor uid
        in: path
        name: vendorId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Get total vendor packs sold
      tags:
      - Analytics
  /analytics/totalRevenue/:vendorId:
    get:
      consumes:
      - application/json
      description: Get the total revenue generated by this vendor
      parameters:
      - description: vendor uid
        in: path
        name: vendorId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Get total revenue
      tags:
      - Analytics
...
  /categories/:category:
    get:
      consumes:
      - application/json
      description: Get list of specific categories
      parameters:
      - description: category param must be item,pack,or vendor
        in: path
        name: category
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Category'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Get list of specific categories
      tags:
      - Category
  /categories/all:
    get:
      consumes:
      - application/json
      description: Get list of all categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Category'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Get list of all categories
      tags:
      - Category
...
swagger: "2.0"